src: Organize static Helpers
